Fix num/double type-safety (clamp returns num), add engine unit tests, fail-safe Pages deploy
- explicit .toDouble()/.toInt() after all clamp() calls (null-safe Dart)
- replace ternary statements with if/else
- add unit tests for ABR engine, buffer monitor, catalog
- job-level continue-on-error for Pages deploy

+++ lib/engine/abr_engine.dart
@@ -90,7 +90,7 @@ class ABREngine {
 
     final target = QualityTier.nearestDown(usable.round());
     var idx = QualityTier.ladder.indexOf(target);
-    if (dataSaver) idx = idx.clamp(0, 2);
+    if (dataSaver) idx = idx.clamp(0, 2).toInt();
 
     // 3) HYSTERESIS — anti-flapping.
     final candidate = QualityTier.ladder[idx];
